requirements.txt
# Core framework
Flask==3.0.3
gevent==24.2.1
gunicorn==22.0.0
requests==2.32.3
python-dotenv==1.0.1

# TensorFlow stack (compatible versions)
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
h5py==3.8.0
pillow==10.3.0
